requirements.txt
librosa==0.6.3
pandas==0.24.2
tqdm==4.31.1
matplotlib==3.0.3
ipython==7.4.0
pyyaml==5.1.0
sed-eval